sched/alt: [Sync] 36c6e17bf169 sched/core: Print out straggler tasks in sched_cpu_dying()

+++ kernel/sched/alt_sched.h
@@ -153,7 +153,7 @@ struct rq {
 	u64 last_ts_switch;
 	u64 clock_task;
 
-	unsigned long nr_running;
+	unsigned int  nr_running;
 	unsigned long nr_uninterruptible;
 
 #ifdef CONFIG_SCHED_HRTICK
